+++ app/Wiki.php
@@ -86,6 +86,10 @@ class Wiki extends Model
     {
         return $this->hasOne(WikiDb::class);
     }
+    public function wikiDeletionReason(): \Illuminate\Database\Eloquent\Relations\HasOne
+    {
+        return $this->hasOne(WikiDeletionReason::class);
+    }
 
     public function wikiSiteStats(): \Illuminate\Database\Eloquent\Relations\HasOne
     {