+++ database/factories/WikiDeletionReasonFactory.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Database\Factories;
+
+use App\WikiDeletionReason;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class WikiDeletionReasonFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = WikiDeletionReason::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [];
+    }
+}